# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: chizlab-admin API
  version: 1.0.0
  description: |
    Educational-materials storage system API.

    **Two route groups:**
    - `/api/admin/*` — full CRUD, JWT-protected. Documented here under the `admin` tag.
    - `/api/public/*` — GET-only, restricted to requests originating from an allowed origin (see `PUBLIC_ALLOWED_ORIGINS`). Documented here under the `public` tag.

    **Internal callback** (`/internal/ai-result`) is documented for developer reference only.
    It is never publicly routed.

    **Separate Swagger UIs:**
    - Admin: `GET /docs/admin`
    - Public: `GET /docs/public`
servers:
  - url: /
    description: Local / Docker

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT access token returned by POST /api/admin/auth/login

    InternalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Shared secret for the AI-worker callback. Validated against INTERNAL_CALLBACK_SECRET env var.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    AiResultCallback:
      type: object
      required:
        - materialId
        - success
      properties:
        materialId:
          type: string
          format: uuid
        success:
          type: boolean
          description: True if AI analysis succeeded; false on terminal failure.
        title:
          type: string
          description: AI-generated title. Present only when success=true.
          nullable: true
        description:
          type: string
          description: AI-generated plain-text description. Present only when success=true.
          nullable: true
        tags:
          type: array
          items:
            type: string
          description: AI-generated tags. Present only when success=true.
          nullable: true
        suggestedCategoryId:
          type: string
          format: uuid
          description: Category UUID suggested by AI. Backend validates before applying.
          nullable: true
        error:
          type: string
          description: Human-readable error message. Present only when success=false.
          nullable: true

    # --- Errors --------------------------------------------------------------
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          description: Human-readable error description.
          example: Validation failed
        error:
          type: string
          description: Short error code / category.
          example: Bad Request
          nullable: true

  # -------------------------------------------------------------------------
  # Reusable parameters
  # -------------------------------------------------------------------------
...
paths:
...
  /internal/ai-result:
    post:
      tags:
        - internal
      summary: AI worker result callback (internal)
      description: |
        **INTERNAL ENDPOINT — never publicly routed.**
        Called by the Python AI worker after media analysis completes (success or terminal failure).

        On success: updates material with AI-filled fields, sets `status=ready`, `isReady=true`.
        On failure: leaves `status=pending`, sets `isReady=false`.

        Protected by `X-Internal-Secret` header validated against `INTERNAL_CALLBACK_SECRET` env var.
      operationId: internalAiResult
      security:
        - InternalSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiResultCallback'
            examples:
              success:
                summary: AI analysis succeeded
                value:
                  materialId: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
                  success: true
                  title: 'Uzbekiston tarixi darsligi'
                  description: 'Qadimgi Uzbekiston tarixi haqidagi darslik'
                  tags:
                    - tarix
                    - uzbekiston
                  suggestedCategoryId: '7fa85f64-5717-4562-b3fc-2c963f66afa9'
              failure:
                summary: AI analysis failed after retries
                value:
                  materialId: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
                  success: false
                  error: 'Gemini API timeout after 3 retries'
      responses:
        '200':
          description: Callback processed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '400':
          description: Invalid payload or material ID not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid X-Internal-Secret header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 401
                message: Invalid internal secret
                error: Unauthorized

# ---------------------------------------------------------------------------
# Tags (for grouping in Swagger UI)
# ---------------------------------------------------------------------------
...